move utils/Svg to shared/Svg
