+++ src/components/Editor/SearchBar.js
@@ -1,7 +1,7 @@
 const React = require("react");
 const { DOM: dom, PropTypes } = React;
 const { findDOMNode } = require("react-dom");
-const Svg = require("../utils/Svg");
+const Svg = require("../shared/Svg");
 const { find, findNext, findPrev, removeOverlay } = require("../../utils/source-search");
 const classnames = require("classnames");
 const { debounce, escapeRegExp } = require("lodash");
