+++ src/components/SourceTabs.js
@@ -14,7 +14,7 @@ const classnames = require("classnames");
 const actions = require("../actions");
 const CloseButton = require("./shared/Button/Close");
 const PaneToggleButton = require("./shared/Button/PaneToggle");
-const Svg = require("./utils/Svg");
+const Svg = require("./shared/Svg");
 const Dropdown = React.createFactory(require("./Dropdown"));
 const { showMenu, buildMenu } = require("../utils/menu");
 const { debounce } = require("lodash");
