+++ src/components/SourcesTree.js
@@ -13,7 +13,7 @@ const {
 } = require("../utils/sources-tree.js");
 const ManagedTree = React.createFactory(require("./utils/ManagedTree"));
 const actions = require("../actions");
-const Svg = require("./utils/Svg");
+const Svg = require("./shared/Svg");
 const { throttle } = require("../utils/utils");
 
 let SourcesTree = React.createClass({
